IEquatable for ComponentReference. Ram now a ushort instead of single byte (to measure megabytes instead of gigabytes). Update pragma warning disables.

+++ server/HackThePlanet/Components/Computer/IProcess.cs
@@ -6,7 +6,7 @@ namespace HackThePlanet
     public interface IProcess : IEntityComponent
     {
         string Command { get; }
-        byte RamUse { get; }
+        ushort RamUse { get; }
         string Status { get; set; }
     }
 }
